+++ RiskOfChaos/EffectDefinitions/World/Spawn/SpawnPots.cs
@@ -1,8 +1,11 @@
 ﻿using HG;
+using RiskOfChaos.ConfigHandling;
 using RiskOfChaos.EffectHandling.EffectClassAttributes;
+using RiskOfChaos.EffectHandling.EffectClassAttributes.Data;
 using RiskOfChaos.EffectHandling.EffectClassAttributes.Methods;
 using RiskOfChaos.Utilities;
 using RiskOfChaos.Utilities.Extensions;
+using RiskOfOptions.OptionConfigs;
 using RoR2;
 using System.Collections;
 using UnityEngine;
@@ -14,7 +17,25 @@ namespace RiskOfChaos.EffectDefinitions.World.Spawn
     [ChaosEffect("spawn_pots")]
     public sealed class SpawnPots : BaseEffect, ICoroutineEffect
     {
-        static readonly GameObject _potPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/ExplosivePotDestructible/ExplosivePotDestructibleBody.prefab").WaitForCompletion();
+        [EffectConfig]
+        static readonly ConfigHolder<int> _potCount =
+            ConfigFactory<int>.CreateConfig("Pot Spawn Count", 50)
+                              .Description("How many pots should be spawned per player")
+                              .OptionConfig(new IntSliderConfig
+                              {
+                                  min = 1,
+                                  max = 100
+                              })
+                              .ValueConstrictor(CommonValueConstrictors.GreaterThanOrEqualTo(1))
+                              .Build();
+
+        static GameObject _potPrefab;
+
+        [SystemInitializer]
+        static void Init()
+        {
+            _potPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/ExplosivePotDestructible/ExplosivePotDestructibleBody.prefab").WaitForCompletion();
+        }
 
         [EffectCanActivate]
         static bool CanActivate()
@@ -29,10 +50,9 @@ namespace RiskOfChaos.EffectDefinitions.World.Spawn
         public IEnumerator OnStartCoroutine()
         {
             const float WAIT_BETWEEN_POT_SPAWNS = 0.1f;
-            const int POT_COUNT = 50;
 
             Vector3 spawnPositionOffset = new Vector3(0f, 10f, 0f);
-            for (int i = 0; i < POT_COUNT; i++)
+            for (int i = 0; i < _potCount.Value; i++)
             {
                 foreach (CharacterBody playerBody in PlayerUtils.GetAllPlayerBodies(true))
                 {
@@ -40,8 +60,7 @@ namespace RiskOfChaos.EffectDefinitions.World.Spawn
                     GameObject pot = GameObject.Instantiate(_potPrefab, playerBody.corePosition + spawnPositionOffset + randomOffset, RNG.RandomRotation());
                     NetworkServer.Spawn(pot);
 
-                    CharacterBody body = pot.GetComponent<CharacterBody>();
-                    if (body)
+                    if (pot.TryGetComponent(out CharacterBody body))
                     {
                         body.AddTimedBuff(RoR2Content.Buffs.HiddenInvincibility, 1f);
                     }